requirements.txt
ai4bharat-transliteration==1.1.3
alabaster==0.7.13
antlr4-python3-runtime==4.8
Babel==2.13.1
bitarray==2.8.3
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Cython==3.0.6
docutils==0.20.1
fairseq==0.12.2
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fsspec==2023.10.0
gevent==23.9.1
greenlet==3.0.1
hydra-core==1.0.7
idna==3.6
imagesize==1.4.1
importlib-metadata==6.8.0
indic-nlp-library==0.92
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lxml==4.9.3
MarkupSafe==2.1.3
mock==5.1.0
Morfessor==2.0.6
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
omegaconf==2.0.6
packaging==23.2
pandas==2.1.3
portalocker==2.8.2
progressbar2==4.2.0
protobuf==4.25.1
pyarrow==14.0.1
pycparser==2.21
pydload==1.0.9
Pygments==2.17.2
python-dateutil==2.8.2
python-utils==3.8.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sacrebleu==2.3.3
sacremoses==0.1.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-argparse==0.4.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
torch==2.1.1
torchaudio==2.1.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urduhack==0.1.4
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
